fixed scripts for scale test
git-svn-id: file:///home/quanpt/svnrep/cde@172 d5ca3f82-63d5-4c06-abdd-b62558c2a7bc

+++ mytest/scale-sim/run.sh
@@ -2,7 +2,10 @@
 
 for l in `cat  cslist.ln.txt`
 do
-  ssh $l kill -9 -1
+  if [ "$l" != "plum" ]
+  then
+    ssh $l kill -9 -1
+  fi
 done
 
 rm -rf cde-package
